rename build_x64.sh to build_deb.sh

+++ build_deb.sh
@@ -3,8 +3,12 @@
 export QT_SELECT=qt6
 #export QT_SELECT=qt5
 
+rm *.buildinfo *.changes *.deb
+
 for package in qiperfc qiperfd qiperftray; do
-	ln -s debian/changelog $package/debian/
+    if [ ! -e $package/debian/changelog ]; then
+        ln -s debian/changelog $package/debian/
+    fi
     cd $package
     #debuild -b -uc -us
     dpkg-buildpackage -b --no-sign
